requirements.txt
Flask==2.2.2
pyotp==2.8.0
cryptography==39.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
requests==2.28.2
pytz==2022.7.1